AdaptiveResonance.jl

The Package Guide provides a tutorial explaining how to get started using Documenter.

Some examples of packages using Documenter can be found on the Examples page.

See the Index for the complete list of documented functions and types.

Manual Outline

Package Guide

Examples

Syntax

Contributing

Library Outline

Public Usage

Developer Documentation

Index

AdaptiveResonance.DDVFA — Method

DDVFA()

Implements a DDVFA learner.

# Examples
```julia-repl
julia> DDVFA()
DDVFA
    opts: opts_DDVFA
    supopts: opts_GNFA
    ...
```

AdaptiveResonance.GNFA — Type

GNFA

Gamma-Normalized Fuzzy ART learner struct

# Examples
```julia-repl
julia> GNFA()
GNFA
    opts: opts_GNFA
    ...
```

AdaptiveResonance.GNFA — Method

GNFA(opts)

Implements a Gamma-Normalized Fuzzy ART learner with specified options.

# Examples
```julia-repl
julia> GNFA(opts)
GNFA
    opts: opts_GNFA
    ...
```

AdaptiveResonance.GNFA — Method

GNFA()

Implements a Gamma-Normalized Fuzzy ART learner.

# Examples
```julia-repl
julia> GNFA()
GNFA
    opts: opts_GNFA
    ...
```

AdaptiveResonance.opts_DDVFA — Type

opts_DDVFA()

Distributed Dual Vigilance Fuzzy ART options struct.

# Examples
```julia-repl
julia> opts_DDVFA()
Initialized opts_DDVFA
```

AdaptiveResonance.opts_GNFA — Type

opts_GNFA()

Gamma-Normalized Fuzzy ART options struct.

# Examples
```julia-repl
julia> opts_GNFA()
Initialized GNFA
```

AdaptiveResonance.activation_match! — Method

activation_match!(art::GNFA, x::Array)

Computes the activationa and match functions of the art module against sample x.

# Examples
```julia-repl
julia> my_GNFA = GNFA()
GNFA
    opts: opts_GNFA
    ...
julia> x, y = load_data()
julia> train!(my_GNFA, x)
julia> x_sample = x[:, 1]
julia> activation_match!(my_GNFA, x_sample)
```

AdaptiveResonance.classify — Method

classify(art::GNFA, x::Array)

Predict categories of 'x' using the GNFA model.

Returns predicted categories 'y_hat'

# Examples
```julia-repl
julia> my_GNFA = GNFA()
GNFA
    opts: opts_GNFA
    ...
julia> x, y = load_data()
julia> train!(my_GNFA, x)
julia> y_hat = classify(my_GNFA, y)
```

AdaptiveResonance.complement_code — Method

complement_code(data)

Normalize the data x to [0, 1] and returns the augmented vector [x, 1 - x].

AdaptiveResonance.doGreet — Method

doGreet(name)

Greets the name given.

Examples

julia> doGreet("Julia")
Hello Julia!

AdaptiveResonance.foo — Method

foo(a, b)

The ubiquitous foo function to verify basic functionality in tests. Returns a + b

Examples

julia> foo(1, 2)
3

AdaptiveResonance.greet — Method

greet()

Print a hello world!

Examples

julia> greet()
Hello World!

AdaptiveResonance.initialize! — Method

initialize!()

Initializes a GNFA learner with an intial sample 'x'

# Examples
```julia-repl
julia> my_GNFA = GNFA()
GNFA
    opts: opts_GNFA
    ...
julia> initialize!(my_GNFA, [1 2 3 4])
```

AdaptiveResonance.my_f — Method

my_f(x, y)

Basic inline function to verify basic functionality in tests. Returns 2 * x + y

Examples

julia> my_f(2, 1)
5

AdaptiveResonance.similarity — Method

similarity(method, F2, field_name, gamma_ref)

Compute the similarity metric depending on method with explicit comparisons for the field name.

AdaptiveResonance.train! — Method

train!(ddvfa, data)

Train the DDVFA model on the data.

AdaptiveResonance.train! — Method

train!()

Trains a GNFA learner with dataset 'x' and optional labels 'y'

# Examples
```julia-repl
julia> my_GNFA = GNFA()
GNFA
    opts: opts_GNFA
    ...
julia> x = load_data()
julia> train!(my_GNFA, x)
```