Developer Index

This page lists the types and functions that are internal to the AdaptiveResonance.jl package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

AdaptiveResonance.activation_match!

AdaptiveResonance.activation_match!

AdaptiveResonance.art_activation

AdaptiveResonance.art_match

AdaptiveResonance.art_match

AdaptiveResonance.average

AdaptiveResonance.basic_activation

AdaptiveResonance.basic_match

AdaptiveResonance.centroid

AdaptiveResonance.choice_by_difference

AdaptiveResonance.color_to_gray

AdaptiveResonance.competition_kernel

AdaptiveResonance.complete

AdaptiveResonance.contrast_insensitive_oriented_filtering

AdaptiveResonance.contrast_normalization

AdaptiveResonance.contrast_sensitive_oriented_filtering

AdaptiveResonance.create_category!

AdaptiveResonance.create_category!

AdaptiveResonance.create_category!

AdaptiveResonance.ddt_x

AdaptiveResonance.ddt_y

AdaptiveResonance.ddt_z

AdaptiveResonance.element_min

AdaptiveResonance.gamma_activation

AdaptiveResonance.gamma_match

AdaptiveResonance.get_data_shape

AdaptiveResonance.get_dim

AdaptiveResonance.get_iterator

AdaptiveResonance.get_n_samples

AdaptiveResonance.get_n_weights

AdaptiveResonance.get_n_weights_vec

AdaptiveResonance.get_sample

AdaptiveResonance.init_classify!

AdaptiveResonance.init_train!

AdaptiveResonance.init_train!

AdaptiveResonance.initialize!

AdaptiveResonance.initialize!

AdaptiveResonance.initialize!

AdaptiveResonance.learn

AdaptiveResonance.learn

AdaptiveResonance.learn

AdaptiveResonance.learn!

AdaptiveResonance.learn!

AdaptiveResonance.learn!

AdaptiveResonance.median

AdaptiveResonance.orientation_competition

AdaptiveResonance.oriented_kernel

AdaptiveResonance.patch_orientation_color

AdaptiveResonance.set_threshold!

AdaptiveResonance.similarity

AdaptiveResonance.single

AdaptiveResonance.stopping_conditions

AdaptiveResonance.stopping_conditions

AdaptiveResonance.stopping_conditions

AdaptiveResonance.stopping_conditions

AdaptiveResonance.surround_kernel

AdaptiveResonance.update_iter

AdaptiveResonance.weighted

Types

AdaptiveResonance.ARTIterator

AdaptiveResonance.ARTMatrix

AdaptiveResonance.ARTVector

Constants

AdaptiveResonance.ACTIVATION_FUNCTIONS_DOCS

AdaptiveResonance.ART_DIM

AdaptiveResonance.ART_SAMPLES

AdaptiveResonance.ART_X_W_ARGS

AdaptiveResonance.MATCH_FUNCTIONS_DOCS

Docs

Documentation for all internal names are listed below.

AdaptiveResonance.ACTIVATION_FUNCTIONS_DOCS — Constant

ACTIVATIONFUNCTIONSDOCS

Description

Common docstring for listing available activation functions.

AdaptiveResonance.ARTIterator — Type

ARTIterator

Description

Acceptable iterators for ART module training and inference

AdaptiveResonance.ART_DIM — Constant

ART_DIM

Description

AdaptiveResonance.jl convention for which 2-D dimension contains the feature dimension.

AdaptiveResonance.ART_SAMPLES — Constant

ART_SAMPLES

Description

AdaptiveResonance.jl convention for which 2-D dimension contains the number of samples.

AdaptiveResonance.ART_X_W_ARGS — Constant

ARTXW_ARGS

Description

Shared arguments string for methods using an ART module, sample 'x', and weight vector 'W'.

AdaptiveResonance.MATCH_FUNCTIONS_DOCS — Constant

MATCHFUNCTIONSDOCS

Description

Common docstring for listing available match functions.

AdaptiveResonance.ARTMatrix — Type

ARTMatrix

Description

The type of matrix used by the AdaptiveResonance.jl package, used to configure matrix growth behavior.

AdaptiveResonance.ARTVector — Type

ARTVector

Description

The type of vector used by the AdaptiveResonance.jl package, used to configure vector growth behvior.

AdaptiveResonance.activation_match! — Method

activation_match!(
    art::DVFA,
    x::AbstractVector{T} where T<:Real
)

Summary

Compute and store the activation and match values for the DVFA module.

Method List / Definition Locations

activation_match!(art, x)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:360.

AdaptiveResonance.activation_match! — Method

activation_match!(
    art::FuzzyART,
    x::AbstractVector{T} where T<:Real
)

Summary

Computes the activation and match functions of the art module against sample x.

Arguments

art::FuzzyART: the FuzzyART module to compute the activation and match values for all weights.

x::RealVector: the sample to compute the activation and match functions against.

Examples

julia> my_FuzzyART = FuzzyART()
FuzzyART
    opts: opts_FuzzyART
    ...
julia> x, y = load_data()
julia> train!(my_FuzzyART, x)
julia> x_sample = x[:, 1]
julia> activation_match!(my_FuzzyART, x_sample)

Method List / Definition Locations

activation_match!(art, x)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:389.

AdaptiveResonance.art_activation — Method

art_activation(
    art::ARTModule,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Evaluates the activation function of the ART/ARTMAP module on the sample 'x' with weight 'W'.

Arguments

art::ARTModule: the ARTModule module.

x::RealVector: the sample to use.

W::RealVector: the weight vector to use.

Method List / Definition Locations

art_activation(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:731.

AdaptiveResonance.art_match — Method

art_match(
    art::ARTModule,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Evaluates the match function of the ART/ARTMAP module on sample 'x' with weight 'W'.

Arguments

art::ARTModule: the ARTModule module.

x::RealVector: the sample to use.

W::RealVector: the weight vector to use.

Method List / Definition Locations

art_match(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:327.

art_match(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:722.

AdaptiveResonance.art_match — Method

art_match(
    art::SFAM,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Returns the match function for the Simple Fuzzy ARTMAP module with weight W and sample x.

Method List / Definition Locations

art_match(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:327.

AdaptiveResonance.average — Method

average(field::AbstractVector{T} where T<:Real) -> Any

Summary

Average linkage DDVFA similarity function.

Arguments

field::RealVector: the DDVFA FuzzyART F2 node field (F2.T or F2.M) to compute the linkage for.

Method List / Definition Locations

average(field)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:476.

AdaptiveResonance.basic_activation — Method

basic_activation(
    art::ARTModule,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Simplified FuzzyARTMAP activation function.

Arguments

art::ARTModule: the ARTModule module.

x::RealVector: the sample to use.

W::RealVector: the weight vector to use.

Method List / Definition Locations

basic_activation(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:683.

AdaptiveResonance.basic_match — Method

basic_match(
    art::ARTModule,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Basic match function.

Arguments

art::ARTModule: the ARTModule module.

x::RealVector: the sample to use.

W::RealVector: the weight vector to use.

Method List / Definition Locations

basic_match(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:674.

AdaptiveResonance.centroid — Method

centroid(
    F2::FuzzyART,
    sample::AbstractVector{T} where T<:Real,
    activation::Bool
) -> Any

Summary

Centroid linkage DDVFA similarity function.

Arguments:

F2::FuzzyART: the DDVFA FuzzyART F2 node to compute the linkage method within.

sample::RealVector: the sample to use for computing the linkage to the F2 module.

activation::Bool: flag to use the activation function. False uses the match function.

Method List / Definition Locations

centroid(F2, sample, activation)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:523.

AdaptiveResonance.choice_by_difference — Method

choice_by_difference(
    art::ARTModule,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Default ARTMAP's choice-by-difference activation function.

Arguments

art::ARTModule: the ARTModule module.

x::RealVector: the sample to use.

W::RealVector: the weight vector to use.

Method List / Definition Locations

choice_by_difference(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:710.

AdaptiveResonance.color_to_gray — Method

color_to_gray(image::Array{T<:AbstractFloat, 3}) -> Matrix

Summary

ARTSCENE Stage 1: Color-to-gray image transformation.

Method List / Definition Locations

color_to_gray(image)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:23.

AdaptiveResonance.competition_kernel — Method

competition_kernel(l::Integer, k::Integer; sign) -> Any

Summary

Competition kernel for ARTSCENE: Stage 5.

Method List / Definition Locations

competition_kernel(l, k; sign)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:195.

AdaptiveResonance.complete — Method

complete(field::AbstractVector{T} where T<:Real) -> Any

Summary

Complete linkage DDVFA similarity function.

Arguments

field::RealVector: the DDVFA FuzzyART F2 node field (F2.T or F2.M) to compute the linkage for.

Method List / Definition Locations

complete(field)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:485.

AdaptiveResonance.contrast_insensitive_oriented_filtering — Method

contrast_insensitive_oriented_filtering(
    y::AbstractArray{T} where T<:Real
) -> Any

Summary

ARTSCENE Stage 4: Contrast-insensitive oriented filtering.

Method List / Definition Locations

contrast_insensitive_oriented_filtering(y)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:182.

AdaptiveResonance.contrast_normalization — Method

contrast_normalization(
    image::AbstractArray{T} where T<:Real
) -> Any

Summary

ARTSCENE Stage 2: Constrast normalization.

Method List / Definition Locations

contrast_normalization(image)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:65.

AdaptiveResonance.contrast_sensitive_oriented_filtering — Method

contrast_sensitive_oriented_filtering(
    image::AbstractArray{T} where T<:Real,
    x::AbstractArray{T} where T<:Real
) -> Any

Summary

ARTSCENE Stage 3: Contrast-sensitive oriented filtering.

Method List / Definition Locations

contrast_sensitive_oriented_filtering(image, x)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:151.

AdaptiveResonance.create_category! — Method

Summary

Creates a category for the ARTModule module, expanding the weights and incrementing the category labels.

Arguments

art::ARTModule: the ARTModule module to add a category to.

x::RealVector: the sample to use for adding a category.

y::Integer: the new label for the new category.

Method List / Definition Locations

create_category!(art, sample, label)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:380.

create_category!(art, x, y; new_cluster)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:232.

create_category!(art, x, y)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:257.

create_category!(art, x, y)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:185.

AdaptiveResonance.create_category! — Method

create_category!(
    art::DDVFA,
    sample::AbstractVector{T} where T<:Real,
    label::Integer
) -> Vector{FuzzyART}

Summary

Create a new category by appending and initializing a new FuzzyART node to F2.

Arguments

art::DDVFA: the DDVFA module to create a new FuzzyART category in.

sample::RealVector: the sample to use for instantiating the new category.

label::Integer: the new label to use for the new category.

Method List / Definition Locations

create_category!(art, sample, label)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:380.

AdaptiveResonance.create_category! — Method

create_category!(
    art::DVFA,
    x::AbstractVector{T} where T<:Real,
    y::Integer;
    new_cluster
) -> Vector{Int64}

Summary

Creates a new category for the DVFA modules.

Arguments

art::DVFA: the DVFA module to add a category to.

x::RealVector: the sample to use for adding a category.

y::Integer: the new label for the new category.

Method List / Definition Locations

create_category!(art, x, y; new_cluster)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:232.

AdaptiveResonance.ddt_x — Method

ddt_x(
    x::AbstractArray{T} where T<:Real,
    image::AbstractArray{T} where T<:Real,
    sigma_s::AbstractArray{T} where T<:Real
) -> SharedArrays.SharedArray{Float64, 3}

Summary

Time rate of change of LGN network (ARTSCENE Stage 2).

Method List / Definition Locations

ddt_x(x, image, sigma_s)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:39.

AdaptiveResonance.ddt_y — Method

ddt_y(
    y::AbstractArray{T} where T<:Real,
    X_plus::AbstractArray{T} where T<:Real,
    X_minus::AbstractArray{T} where T<:Real,
    alpha::Real
) -> SharedArrays.SharedArray{Float64, 4}

Summary

Shunting equation for ARTSCENE Stage 3.

Method List / Definition Locations

ddt_y(y, X_plus, X_minus, alpha)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:112.

AdaptiveResonance.ddt_z — Method

ddt_z(
    z::AbstractArray{T} where T<:Real
) -> SharedArrays.SharedArray{Float64, 4}

Summary

Time rate of change for ARTSCENE: Stage 5.

Method List / Definition Locations

ddt_z(z)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:211.

AdaptiveResonance.element_min — Method

element_min(
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Returns the element-wise minimum between sample x and weight W.

Arguments

x::RealVector: the input sample.

W::RealVector: the weight vector to compare the sample against.

Method List / Definition Locations

element_min(x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:222.

AdaptiveResonance.gamma_activation — Method

gamma_activation(
    art::ARTModule,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Gamma-normalized activation funtion.

Arguments

art::ARTModule: the ARTModule module.

x::RealVector: the sample to use.

W::RealVector: the weight vector to use.

Method List / Definition Locations

gamma_activation(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:701.

AdaptiveResonance.gamma_match — Method

gamma_match(
    art::ARTModule,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Gamma-normalized match function.

Arguments

art::ARTModule: the ARTModule module.

x::RealVector: the sample to use.

W::RealVector: the weight vector to use.

Method List / Definition Locations

gamma_match(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:692.

AdaptiveResonance.get_data_shape — Method

get_data_shape(
    data::AbstractMatrix{T} where T<:Real
) -> Tuple{Any, Any}

Summary

Returns the (dim, n_samples) of the provided 2-D data matrix, enforcing the ART package convention.

Arguments

data::RealMatrix: the 2-D data to infer the feature dimension and number of samples from.

Method List / Definition Locations

get_data_shape(data)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:283.

AdaptiveResonance.get_dim — Method

get_dim(data::AbstractMatrix{T} where T<:Real) -> Any

Summary

Returns the dimension of the data, enforcint the (dim, n_samples) convention of the package.

Arguments

data::RealMatrix: the 2-D data to infer the feature dimension of.

Method List / Definition Locations

get_dim(data)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:261.

AdaptiveResonance.get_iterator — Method

get_iterator(opts::ARTOpts, n_samples::Integer) -> Any

Summary

Creates an iterator object according to the ART/ARTMAP modules display settings for batch iteration.

Arguments

opts::ARTOpts: the ART/ARTMAP module's options containing display settings.

n_samples::Integer: the number of iterations to create the iterator for.

Method List / Definition Locations

get_iterator(opts, n_samples)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:443.

AdaptiveResonance.get_n_samples — Method

get_n_samples(data::AbstractMatrix{T} where T<:Real) -> Any

Summary

Returns the number of samples, enforcing the convention of the package.

Arguments

data::RealMatrix: the 2-D data to infer the number of samples from.

Method List / Definition Locations

get_n_samples(data)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:272.

AdaptiveResonance.get_n_weights — Method

get_n_weights(art::DDVFA) -> Int64

Summary

Convenience function; return the sum total number of weights in the DDVFA module.

Method List / Definition Locations

get_n_weights(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:564.

AdaptiveResonance.get_n_weights_vec — Method

get_n_weights_vec(art::DDVFA) -> Vector{Int64}

Summary

Convenience function; return the number of weights in each category as a vector.

Arguments

art::DDVFA: the DDVFA module to get all of the weights from as a list.

Method List / Definition Locations

get_n_weights_vec(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:557.

AdaptiveResonance.get_sample — Method

get_sample(
    x::AbstractMatrix{T} where T<:Real,
    i::Integer
) -> Any

Summary

Returns a sample from data array x at sample location i. This function implements the convention that columns are samples while rows are features within samples.

Arguments

x::RealMatrix: the batch of data to grab a sample from.

i::Integer: the index to get the sample from.

Method List / Definition Locations

get_sample(x, i)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:478.

AdaptiveResonance.init_classify! — Method

init_classify!(
    x::AbstractArray{T} where T<:Real,
    art::ARTModule,
    preprocessed::Bool
) -> Any

Summary

Initializes the classification loop for batch inference.

Arguments

x::RealArray: the data that is used for inference.

art::ARTModule: the ART/ARTMAP module that will be used for inference.

preprocessed::Bool: required flag for if the data has already been complement coded and normalized.

Method List / Definition Locations

init_classify!(x, art, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:543.

AdaptiveResonance.init_train! — Method

init_train!(
    x::AbstractMatrix{T} where T<:Real,
    art::ARTModule,
    preprocessed::Bool
) -> Any

Summary

Initializes the training loop for batch learning.

Arguments

x::RealMatrix: the data that is used for training.

art::ARTModule: the ART/ARTMAP that will be trained.

preprocessed::Bool: required flag for if the data has already been complement coded and normalized.

Method List / Definition Locations

init_train!(x, art, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:525.

AdaptiveResonance.init_train! — Method

init_train!(
    x::AbstractVector{T} where T<:Real,
    art::ARTModule,
    preprocessed::Bool
) -> AbstractVector{T} where T<:Real

Summary

Initializes the module for training in a single iteration.

The purpose of this function is mainly to handle the conditions of complement coding. Fails if the module was incorrectly set up or if the module was not setup and the data was not preprocessed.

Arguments

x::RealVector: the sample used for initialization.

art::ARTModule: the ART/ARTMAP module that will be trained on the sample.

preprocessed::Bool: a required flag for if the sample has already been complement coded and normalized.

Method List / Definition Locations

init_train!(x, art, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:494.

AdaptiveResonance.initialize! — Method

Summary

Initializes the ART module for training with sample 'x' and optional label 'y', setting up the data configuration and instantiating the first category.

This function is used during the first training iteration when the ART module is empty.

Arguments

art::ART: the ART module to initialize.

x::RealVector: the sample to use for initialization.

y::Integer=0: the optional new label for the first weight of the ART module. If not specified, defaults the new label to 1.

Examples

julia> my_FuzzyART = FuzzyART()
FuzzyART
    opts: opts_FuzzyART
    ...
julia> initialize!(my_FuzzyART, [1, 2, 3, 4])

# Method List / Definition Locations

julia initialize!(art, x; y)

defined at [`/home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:213`](https://github.com/AP6YC/AdaptiveResonance.jl/tree/d0a55ab955914bd72b9f7e35a74af858cbae2f7f//src/ART/DVFA.jl#L213).

julia initialize!(art, x; y) ```

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:242.

AdaptiveResonance.initialize! — Method

Summary

Initializes the supervised ARTMAP module for training with sample 'x' and label 'y', setting up the data configuration and instantiating the first category.

Arguments

art::ARTMAP: the ARTMAP module to initialize.

x::RealVector: the sample to use for initialization.

y::Integer: the initial supervised label.

Examples

julia> my_sfam = SFAM()
SFAM
    opts: opts_SFAM
    ...
julia> initialize!(my_SFAM, [1, 2, 3, 4])

# Method List / Definition Locations

julia initialize!(art, x, y) ```

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:177.

AdaptiveResonance.initialize! — Method

initialize!(
    art::DVFA,
    x::AbstractVector{T} where T<:Real;
    y
) -> Vector{Int64}

Summary

Initializes a DVFA learner with an initial sample 'x'.

This function is used during the first training iteraction when the DVFA module is empty.

Arguments

art::DVFA: the DVFA module to initialize.

x::RealVector: the sample to use for initialization.

y::Integer=0: the optional new label for the first weight of the FuzzyART module. If not specified, defaults the new label to 1.

Method List / Definition Locations

initialize!(art, x; y)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:213.

AdaptiveResonance.learn! — Method

learn!(
    art::DVFA,
    x::AbstractVector{T} where T<:Real,
    index::Integer
) -> Any

Summary

In place learning function.

Method List / Definition Locations

learn!(art, x, index)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:381.

AdaptiveResonance.learn! — Method

learn!(
    art::FuzzyART,
    x::AbstractVector{T} where T<:Real,
    index::Integer
) -> Int64

Summary

In place learning function with instance counting.

Arguments

art::FuzzyART: the FuzzyART module to update.

x::RealVector: the sample to learn from.

index::Integer: the index of the FuzzyART weight to update.

Method List / Definition Locations

learn!(art, x, index)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:442.

AdaptiveResonance.learn! — Method

learn!(
    art::SFAM,
    x::AbstractVector{T} where T<:Real,
    index::Integer
) -> Any

Summary

In-place learning function.

Method List / Definition Locations

learn!(art, x, index)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:318.

AdaptiveResonance.learn — Method

learn(
    art::DVFA,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Return the modified weight of the DVFA module conditioned by sample x.

Method List / Definition Locations

learn(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:373.

AdaptiveResonance.learn — Method

learn(
    art::FuzzyART,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Return the modified weight of the art module conditioned by sample x.

Arguments

art::FuzzyART: the FuzzyART module containing learning options.

x::RealVector: the sample to learn from.

W::RealVector: the weight vector to update against the sample.

Method List / Definition Locations

learn(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:429.

AdaptiveResonance.learn — Method

learn(
    art::SFAM,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Returns a single updated weight for the Simple Fuzzy ARTMAP module for weight vector W and sample x.

Method List / Definition Locations

learn(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:310.

AdaptiveResonance.median — Method

median(field::AbstractVector{T} where T<:Real) -> Any

Summary

Median linkage DDVFA similarity function.

Arguments

field::RealVector: the DDVFA FuzzyART F2 node field (F2.T or F2.M) to compute the linkage for.

Method List / Definition Locations

median(field)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:494.

AdaptiveResonance.orientation_competition — Method

orientation_competition(
    z::AbstractArray{T} where T<:Real
) -> Any

Summary

ARTSCENE Stage 5: Orientation competition at the same position.

Method List / Definition Locations

orientation_competition(z)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:236.

AdaptiveResonance.oriented_kernel — Method

oriented_kernel(
    i::Integer,
    j::Integer,
    p::Integer,
    q::Integer,
    k::Integer,
    sigma_h::Real,
    sigma_v::Real;
    sign
) -> Any

Summary

Oriented, elongated, spatially offset kernel G for ARTSCENE Stage 3.

Method List / Definition Locations

oriented_kernel(i, j, p, q, k, sigma_h, sigma_v; sign)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:90.

AdaptiveResonance.patch_orientation_color — Method

patch_orientation_color(
    z::AbstractArray{T} where T<:Real,
    image::AbstractArray{T} where T<:Real
) -> Tuple{Union{Array{Float64, 4}, OffsetArrays.OffsetArray{Float64, 4, Array{Float64, 4}}}, Array{Float64, 3}}

Summary

ARTSCENE Stage 6: Create patch feature vectors.

Method List / Definition Locations

patch_orientation_color(z, image)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:257.

AdaptiveResonance.set_threshold! — Method

Summary

Sets the match threshold of the ART/ARTMAP module as a function of the vigilance parameter.

Depending on selected ART/ARTMAP module and its options, this may be a function of other parameters as well.

Arguments

art::ARTModule: the ART/ARTMAP module for setting a new threshold.

Method List / Definition Locations

set_threshold!(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:240.

set_threshold!(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:197.

set_threshold!(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:233.

AdaptiveResonance.similarity — Method

similarity(
    method::Symbol,
    F2::FuzzyART,
    sample::AbstractVector{T} where T<:Real,
    activation::Bool
) -> Any

Summary

Compute the similarity metric depending on method with explicit comparisons for the field name.

Arguments

method::Symbol: the linkage method to use.

F2::FuzzyART: the DDVFA FuzzyART F2 node to compute the linkage method within.

sample::RealVector: the sample to use for computing the linkage to the F2 module.

activation::Bool: flag to use the activation function. False uses the match function.

Method List / Definition Locations

similarity(method, F2, sample, activation)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:435.

AdaptiveResonance.single — Method

single(field::AbstractVector{T} where T<:Real) -> Any

Summary

Single linkage DDVFA similarity function.

Arguments

field::RealVector: the DDVFA FuzzyART F2 node field (F2.T or F2.M) to compute the linkage for.

Method List / Definition Locations

single(field)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:467.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::DDVFA) -> Bool

Summary

Stopping conditions for Distributed Dual Vigilance Fuzzy ARTMAP.

Returns true if there is no change in weights during the epoch or the maxmimum epochs has been reached.

Arguments

art::DDVFA: the DDVFA module for checking stopping conditions.

Method List / Definition Locations

stopping_conditions(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:396.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::DVFA) -> Bool

Summary

Stopping conditions for a DVFA module.

Method List / Definition Locations

stopping_conditions(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:389.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::FuzzyART) -> Bool

Summary

Stopping conditions for a FuzzyART module.

Arguments

art::FuzzyART: the FuzzyART module to check stopping conditions for.

Method List / Definition Locations

stopping_conditions(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:454.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::SFAM) -> Bool

Summary

Stopping conditions for Simple Fuzzy ARTMAP, checked at the end of every epoch.

Method List / Definition Locations

stopping_conditions(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:301.

AdaptiveResonance.surround_kernel — Method

surround_kernel(
    i::Integer,
    j::Integer,
    p::Integer,
    q::Integer,
    scale::Integer
) -> Any

Summary

Surround kernel S function for ARTSCENE Stage 2.

Method List / Definition Locations

surround_kernel(i, j, p, q, scale)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:32.

AdaptiveResonance.update_iter — Method

update_iter(
    art::ARTModule,
    iter::Union{ProgressBars.ProgressBar, UnitRange},
    i::Integer
) -> Union{Nothing, String}

Summary

Updates the iteration of the ART/ARTMAP module, training or inference, according to its display settings.

Arguments

art::ARTModule: the ART/ARTMAP module being iterated upon.

iter::ARTIterator: the iterator object used in the training/inference loop.

i::Integer: the iteration during training/inference that the iterator should be updated to.

Method List / Definition Locations

update_iter(art, iter, i)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:461.

AdaptiveResonance.weighted — Method

weighted(F2::FuzzyART, activation::Bool) -> Float64

Summary

Weighted linkage DDVFA similarity function.

Arguments:

F2::FuzzyART: the DDVFA FuzzyART F2 node to compute the linkage method within.

activation::Bool: flag to use the activation function. False uses the match function.

Method List / Definition Locations

weighted(F2, activation)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:505.