Classification and Evaluation

Classify a batch or an individual sample, then compare predictions with known labels. See Mismatch vs. Best-Matching-Unit for get_bmu behavior.

Index

classify

performance

Classification

classify

AdaptiveResonance.classify — Function

classify(
    art::ARTModule,
    x::AbstractMatrix{T} where T<:Real;
    preprocessed,
    get_bmu
) -> Any

Summary

Predict categories of 'x' using the ART model.

Returns predicted categories 'y_hat.'

Arguments

art::ARTModule: ART or ARTMAP module to use for batch inference.

x::RealMatrix: the 2-D dataset containing columns of samples with rows of features.

preprocessed::Bool=false: flag, if the data has already been complement coded or not.

get_bmu::Bool=false, flag, if the model should return the best-matching-unit label in the case of total mismatch.

Examples

julia> my_DDVFA = DDVFA()
DDVFA
    opts: opts_DDVFA
    ...
julia> x, y = load_data()
julia> train!(my_DDVFA, x)
julia> y_hat = classify(my_DDVFA, y)

Method List / Definition Locations

classify(art, x; preprocessed, get_bmu)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/lib/common.jl:554.

Summary

Predict categories of a single sample of features 'x' using the ART model.

Returns predicted category 'y_hat.'

Arguments

art::ARTModule: ART or ARTMAP module to use for batch inference.

x::RealVector: the single sample of features to classify.

preprocessed::Bool=false: optional, flag if the data has already been complement coded or not.

get_bmu::Bool=false: optional, flag if the model should return the best-matching-unit label in the case of total mismatch.

Method List / Definition Locations

classify(art, x; preprocessed, get_bmu)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/distributed/modules/DDVFA.jl:362.

classify(art, x; preprocessed, get_bmu)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/single/modules/DVFA.jl:360.

classify(art, x; preprocessed, get_bmu)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/single/modules/FuzzyART.jl:392.

classify(art, x; preprocessed, get_bmu)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/single/modules/HypersphereART.jl:265.

classify(art, x; preprocessed, get_bmu)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:315.

Evaluation

performance

AdaptiveResonance.performance — Function

performance(
    y_hat::AbstractVector{T} where T<:Integer,
    y::AbstractVector{T} where T<:Integer
) -> Any

Summary

Convenience function to get the categorization performance of y_hat against y.

Arguments

y_hat::IntegerVector: the estimated labels.

y::IntegerVector: the true labels.

Method List / Definition Locations

performance(y_hat, y)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/lib/common.jl:200.