ARTMAP Models and Options

This page lists ARTMAP models that use supervisory labels. See ART vs. ARTMAP to learn the difference between them.

Index

SFAM, opts_SFAM

FAM, opts_FAM

DAM, opts_DAM

SFAM

SFAM

AdaptiveResonance.SFAM — Type

mutable struct SFAM <: ARTMAP

Summary

Simple Fuzzy ARTMAP struct.

For module options, see AdaptiveResonance.opts_SFAM.

References

G. A. Carpenter, S. Grossberg, N. Markuzon, J. H. Reynolds, and D. B. Rosen, “Fuzzy ARTMAP: A Neural Network Architecture for Incremental Supervised Learning of Analog Multidimensional Maps,” IEEE Trans. Neural Networks, vol. 3, no. 5, pp. 698-713, 1992, doi: 10.1109/72.159059.

Fields

opts::opts_SFAM: Simplified Fuzzy ARTMAP options struct.

config::DataConfig: Data configuration struct.

W::ElasticArrays.ElasticMatrix{Float64, V} where V<:DenseVector{Float64}: Category weight matrix.

labels::Vector{Int64}: Incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

n_categories::Int64: Number of category weights (F2 nodes).

epoch::Int64: Current training epoch.

T::Vector{Float64}: DDVFA activation values.

M::Vector{Float64}: DDVFA match values.

stats::Dict{String, Any}: Runtime statistics for the module, implemented as a dictionary containing entries at the end of each training iteration. These entries include the best-matching unit index and the activation and match values of the winning node.

opts_SFAM

AdaptiveResonance.opts_SFAM — Type

mutable struct opts_SFAM <: ARTOpts

Summary

Implements a Simple Fuzzy ARTMAP learner's options.

These options are a Parameters.jl struct, taking custom options keyword arguments. Each field has a default value listed below.

Fields

rho::Float64: Vigilance parameter: rho ∈ [0, 1]. Default: 0.75

alpha::Float64: Choice parameter: alpha > 0. Default: 1.0e-7

epsilon::Float64: Match tracking parameter: epsilon ∈ (0, 1). Default: 0.001

beta::Float64: Learning parameter: beta ∈ (0, 1]. Default: 1.0

max_epoch::Int64: Maximum number of epochs during training: max_epoch ∈ [1, Inf). Default: 1

display::Bool: Display flag for progress bars. Default: false

uncommitted::Bool: Flag to use an uncommitted node when learning.

If true, new weights are created with ones(dim) and learn on the complement-coded sample. If false, fast-committing is used where the new weight is simply the complement-coded sample. Default: false

match::Symbol: Selected match function. Default: :basic_match

activation::Symbol: Selected activation function. Default: :basic_activation

update::Symbol: Selected weight update function. Default: :basic_update

sort::Bool: Flag to sort the F2 nodes by activation before the match phase

When true, the F2 nodes are sorted by activation before match. When false, an iterative argmax and inhibition procedure is used to find the best-matching unit. Default: false

FAM

FAM

AdaptiveResonance.FAM — Type

mutable struct FAM <: ARTMAP

Summary

Fuzzy ARTMAP struct.

For module options, see AdaptiveResonance.opts_FAM.

References

G. A. Carpenter, S. Grossberg, N. Markuzon, J. H. Reynolds, and D. B. Rosen, “Fuzzy ARTMAP: A Neural Network Architecture for Incremental Supervised Learning of Analog Multidimensional Maps,” IEEE Trans. Neural Networks, vol. 3, no. 5, pp. 698-713, 1992, doi: 10.1109/72.159059.

Fields

opts::opts_FAM: Fuzzy ARTMAP options struct.

config::DataConfig: Data configuration struct.

W::ElasticArrays.ElasticMatrix{Float64, V} where V<:DenseVector{Float64}: Category weight matrix.

labels::Vector{Int64}: Incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

n_categories::Int64: Number of category weights (F2 nodes).

epoch::Int64: Current training epoch.

opts_FAM

AdaptiveResonance.opts_FAM — Type

mutable struct opts_FAM <: ARTOpts

Summary

Implements a Fuzzy ARTMAP learner's options.

These options are a Parameters.jl struct, taking custom options keyword arguments. Each field has a default value listed below.

Fields

rho::Float64: Vigilance parameter: rho ∈ [0, 1]. Default: 0.6

alpha::Float64: Choice parameter: alpha > 0. Default: 1.0e-7

epsilon::Float64: Match tracking parameter: epsilon ∈ (0, 1). Default: 0.001

beta::Float64: Learning parameter: beta ∈ (0, 1]. Default: 1.0

max_epochs::Int64: Maximum number of epochs during training: max_epochs ∈ [1, Inf) Default: 1

uncommitted::Bool: Uncommitted node flag. Default: true

display::Bool: Display flag for progress bars. Default: false

sort::Bool: Flag to sort the F2 nodes by activation before the match phase

When true, the F2 nodes are sorted by activation before match. When false, an iterative argmax and inhibition procedure is used to find the best-matching unit. Default: false

DAM

DAM

AdaptiveResonance.DAM — Function

DAM(; kwargs...) -> SFAM

Summary

Constructs a Default ARTMAP module using a SFAM module using Default ARTMAP's choice-by-difference activation function.

Default ARTMAP is a variant of SFAM, using the AdaptiveResonance.opts_SFAM options. This constructor sets the activation to :choice_by_difference in addition to the keyword argument options you provide.

Arguments

kwargs: keyword arguments of Simplified FuzzyARTMAP options (see AdaptiveResonance.opts_SFAM)

References:

G. P. Amis and G. A. Carpenter, 'Default ARTMAP 2,' IEEE Int. Conf. Neural Networks - Conf. Proc., vol. 2, no. September 2007, pp. 777-782, Mar. 2007, doi: 10.1109/IJCNN.2007.4371056.

Method List / Definition Locations

DAM(; kwargs...)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/variants.jl:29.

DAM(opts::opts_SFAM) -> SFAM

Summary

Implements a Default ARTMAP module with specified options.

Default ARTMAP is a variant of SFAM, using the AdaptiveResonance.opts_SFAM options. This constructor sets the activation to :choice_by_difference in addition to the keyword argument options you provide.

Arguments

opts::opts_SFAM: the Simplified FuzzyARTMAP options (see AdaptiveResonance.opts_SFAM).

Method List / Definition Locations

DAM(opts)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/variants.jl:41.

opts_DAM

AdaptiveResonance.opts_DAM — Function

opts_DAM(; kwargs...) -> opts_SFAM

Summary

Implements a Default ARTMAP module's options.

Default ARTMAP is a variant of SFAM, using the AdaptiveResonance.opts_SFAM options. This constructor sets the activation to :choice_by_difference in addition to the keyword argument options you provide.

These options are a Parameters.jl struct, taking custom options keyword arguments. Each field has a default value listed below.

Method List / Definition Locations

opts_DAM(; kwargs...)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/variants.jl:52.