Training

Use a feature matrix for batch training or a feature vector for incremental training. ART accepts optional labels with y; ARTMAP takes labels as a positional argument. The documented dispatch variants are collected below.

Index

train!

Training Models

train!

AdaptiveResonance.train! — Function

train!(
    art::ARTMAP,
    x::AbstractMatrix{T} where T<:Real,
    y::AbstractVector{T} where T<:Integer
) -> Any
train!(
    art::ARTMAP,
    x::AbstractMatrix{T} where T<:Real,
    y::AbstractVector{T} where T<:Integer,
    preprocessed::Bool
) -> Any

Summary

train!(art::ARTMAP, x::RealMatrix, y::IntegerVector, preprocessed::Bool=false)

Train the ARTMAP model on a batch of data 'x' with supervisory labels 'y.'

Arguments

art::ARTMAP: the supervised ARTMAP model to train.

x::RealMatrix: the 2-D dataset containing columns of samples with rows of features.

y::IntegerVector: labels for supervisory training.

preprocessed::Bool=false: flag, if the data has already been complement coded or not.

Method List / Definition Locations

train!(art, x, y)
train!(art, x, y, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/common.jl:23.

Summary

Train the supervised ARTMAP model on a single sample of features 'x' with supervisory label 'y'.

Arguments

art::ARTMAP: the supervised ART model to train.

x::RealVector: the single sample feature vector to train upon.

y::Integer: the label for supervisory training.

preprocessed::Bool=false: optional, flag if the data has already been complement coded or not.

Method List / Definition Locations

train!(art, x, y; preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:242.

train!(
    art::ART,
    x::AbstractMatrix{T} where T<:Real;
    y,
    preprocessed
) -> Any

Summary

Train the ART model on a batch of data 'x' with optional supervisory labels 'y.'

Arguments

art::ART: the unsupervised ART model to train.

x::RealMatrix: the 2-D dataset containing columns of samples with rows of features.

y::IntegerVector=Int[]: optional, labels for simple supervisory training.

preprocessed::Bool=false: optional, flag if the data has already been complement coded or not.

Method List / Definition Locations

train!(art, x; y, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/common.jl:21.

Summary

Train the ART model on a single sample of features 'x' with an optional supervisory label.

Arguments

art::ART: the unsupervised ART model to train.

x::RealVector: the single sample feature vector to train upon.

y::Integer=0: optional, a label for simple supervisory training.

preprocessed::Bool=false: optional, flag if the data has already been complement coded or not.

Method List / Definition Locations

train!(art, x; y, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/distributed/modules/DDVFA.jl:285.

train!(art, x; y, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/single/modules/DVFA.jl:278.

train!(art, x; y, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/single/modules/FuzzyART.jl:310.

train!(art, x; y, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/single/modules/HypersphereART.jl:216.