AdaptiveResonance.jl

The Package Guide provides a tutorial explaining how to get started using Documenter.

Some examples of packages using Documenter can be found on the Examples page.

See the Index for the complete list of documented functions and types.

Manual Outline

Package Guide

Examples

Syntax

Contributing

Library Outline

Public Usage

Developer Documentation

Index

AdaptiveResonance.DAM — Type

DAM

Default ARTMAP struct.

AdaptiveResonance.DAM — Method

DAM(opts)

Implements a Default ARTMAP learner with specified options

Examples

julia> opts = opts_DAM()
julia> DAM(opts)
DAM
    opts: opts_DAM
    ...

AdaptiveResonance.DAM — Method

DAM()

Implements a Default ARTMAP learner.

Examples

julia> DAM()
DAM
    opts: opts_DAM
    ...

AdaptiveResonance.DDVFA — Type

DDVFA

Distributed Dual Vigilance Fuzzy ARTMAP module struct.

Examples

julia> DDVFA()
DDVFA
    opts: opts_DDVFA
    supopts::opts_GNFA
    ...

AdaptiveResonance.DDVFA — Method

DDVFA(opts::opts_DDVFA)

Implements a DDVFA learner with specified options.

Examples

julia> my_opts = opts_DDVFA()
julia> DDVFA(my_opts)
DDVFA
    opts: opts_DDVFA
    supopts: opts_GNFA
    ...

AdaptiveResonance.DDVFA — Method

DDVFA()

Implements a DDVFA learner.

Examples

julia> DDVFA()
DDVFA
    opts: opts_DDVFA
    supopts: opts_GNFA
    ...

AdaptiveResonance.FAM — Type

FAM

Fuzzy ARTMAP struct.

AdaptiveResonance.FAM — Method

FAM(opts)

Implements a Fuzzy ARTMAP learner with specified options.

Examples

julia> opts = opts_FAM()
julia> FAM(opts)
FAM
    opts: opts_FAM
    ...

AdaptiveResonance.FAM — Method

FAM()

Implements a Fuzzy ARTMAP learner.

Examples

julia> FAM()
FAM
    opts: opts_FAM
    ...

AdaptiveResonance.GNFA — Type

GNFA

Gamma-Normalized Fuzzy ART learner struct

Examples

julia> GNFA()
GNFA
    opts: opts_GNFA
    ...

AdaptiveResonance.GNFA — Method

GNFA(opts)

Implements a Gamma-Normalized Fuzzy ART learner with specified options.

Examples

julia> GNFA(opts)
GNFA
    opts: opts_GNFA
    ...

AdaptiveResonance.GNFA — Method

GNFA()

Implements a Gamma-Normalized Fuzzy ART learner.

Examples

julia> GNFA()
GNFA
    opts: opts_GNFA
    ...

AdaptiveResonance.SFAM — Type

SFAM

Simple Fuzzy ARTMAP struct.

AdaptiveResonance.SFAM — Method

SFAM(opts)

Implements a Simple Fuzzy ARTMAP learner with specified options.

Examples

julia> opts = opts_SFAM()
julia> SFAM(opts)
SFAM
    opts: opts_SFAM
    ...

AdaptiveResonance.SFAM — Method

SFAM()

Implements a Simple Fuzzy ARTMAP learner.

Examples

julia> SFAM()
SFAM
    opts: opts_SFAM
    ...

AdaptiveResonance.opts_DAM — Type

opts_DAM()

Implements a Default ARTMAP learner's options.

Examples

julia> my_opts = opts_DAM()

AdaptiveResonance.opts_DDVFA — Type

opts_DDVFA()

Distributed Dual Vigilance Fuzzy ART options struct.

Examples

julia> opts_DDVFA()
Initialized opts_DDVFA

AdaptiveResonance.opts_FAM — Type

opts_FAM()

Implements a Fuzzy ARTMAP learner's options.

Examples

julia> my_opts = opts_FAM()

AdaptiveResonance.opts_GNFA — Type

opts_GNFA()

Gamma-Normalized Fuzzy ART options struct.

Examples

julia> opts_GNFA()
Initialized GNFA

AdaptiveResonance.opts_SFAM — Type

opts_SFAM()

Implements a Simple Fuzzy ARTMAP learner's options.

Examples

julia> my_opts = opts_SFAM()

AdaptiveResonance.activation — Method

activation(art::DAM, x, W)

Default ARTMAP's choice-by-difference activation function.

AdaptiveResonance.activation — Method

activation(art::SFAM, x, W)

Returns the activation value of the Simple Fuzzy ARTMAP module with weight W and sample x.

AdaptiveResonance.activation_match! — Method

activation_match!(art::GNFA, x::Array)

Computes the activationa and match functions of the art module against sample x.

Examples

julia> my_GNFA = GNFA()
GNFA
    opts: opts_GNFA
    ...
julia> x, y = load_data()
julia> train!(my_GNFA, x)
julia> x_sample = x[:, 1]
julia> activation_match!(my_GNFA, x_sample)

AdaptiveResonance.art_match — Method

art_match(art::DAM, x, W)

Returns the match function for the Default ARTMAP module with weight W and sample x.

AdaptiveResonance.art_match — Method

art_match(art::SFAM, x, W)

Returns the match function for the Simple Fuzzy ARTMAP module with weight W and sample x.

AdaptiveResonance.classify — Method

classify(art, x)

Categorize data 'x' using a trained Default ARTMAP module 'art'.

Examples

julia> x, y = load_data()
julia> x_test, y_test = load_test_data()
julia> art = DAM()
DAM
    opts: opts_DAM
    ...
julia> train!(art, x, y)
julia> classify(art, x_test)

AdaptiveResonance.classify — Method

classify(art::GNFA, x::Array)

Predict categories of 'x' using the GNFA model.

Returns predicted categories 'y_hat'

Examples

julia> my_GNFA = GNFA()
GNFA
    opts: opts_GNFA
    ...
julia> x, y = load_data()
julia> train!(my_GNFA, x)
julia> y_hat = classify(my_GNFA, y)

AdaptiveResonance.classify — Method

classify(art, x)

Categorize data 'x' using a trained Simple Fuzzy ARTMAP module 'art'.

Examples

julia> x, y = load_data()
julia> x_test, y_test = load_test_data()
julia> art = SFAM()
SFAM
    opts: opts_SFAM
    ...
julia> train!(art, x, y)
julia> classify(art, x_test)

AdaptiveResonance.complement_code — Method

complement_code(data)

Normalize the data x to [0, 1] and returns the augmented vector [x, 1 - x].

AdaptiveResonance.element_min — Method

element_min(x::Array, W::Array)

Returns the element-wise minimum between sample x and weight W.

AdaptiveResonance.initialize! — Method

initialize!()

Initializes a GNFA learner with an intial sample 'x'

Examples

julia> my_GNFA = GNFA()
GNFA
    opts: opts_GNFA
    ...
julia> initialize!(my_GNFA, [1 2 3 4])

AdaptiveResonance.learn! — Method

learn(art::GNFA, x::Array, W::Array)

In place learning function with instance counting.

AdaptiveResonance.learn — Method

learn(art::DAM, x, W)

Returns a single updated weight for the Simple Fuzzy ARTMAP module for weight vector W and sample x.

AdaptiveResonance.learn — Method

learn(art::GNFA, x::Array, W::Array)

Return the modified weight of the art module conditioned by sample x.

AdaptiveResonance.learn — Method

learn(art::SFAM, x, W)

Returns a single updated weight for the Simple Fuzzy ARTMAP module for weight vector W and sample x.

AdaptiveResonance.performance — Method

performance(y_hat, y)

Returns the categorization performance of y_hat against y.

AdaptiveResonance.similarity — Method

similarity(method, F2, field_name, gamma_ref)

Compute the similarity metric depending on method with explicit comparisons for the field name.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::DAM)

Stopping conditions for Default ARTMAP, checked at the end of every epoch.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::DDVFA)

Stopping conditions for Distributed Dual Vigilance Fuzzy ARTMAP. Returns true if there is no change in weights during the epoch or the maxmimum epochs has been reached.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::GNFA)

Stopping conditions for a GNFA module.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::SFAM)

Stopping conditions for Simple Fuzzy ARTMAP, checked at the end of every epoch.

AdaptiveResonance.train! — Method

train(art::DAM, x, y)

Trains a Default ARTMAP learner in a supervised manner.

Examples

julia> x, y = load_data()
julia> art = DAM()
DAM
    opts: opts_DAM
    ...
julia> train!(art, x, y)

AdaptiveResonance.train! — Method

train!(ddvfa, data)

Train the DDVFA model on the data.

AdaptiveResonance.train! — Method

train!()

Trains a GNFA learner with dataset 'x' and optional labels 'y'

Examples

julia> my_GNFA = GNFA()
GNFA
    opts: opts_GNFA
    ...
julia> x = load_data()
julia> train!(my_GNFA, x)

AdaptiveResonance.train! — Method

train(art, x, y)

Trains a Simple Fuzzy ARTMAP learner in a supervised manner.

Examples

julia> x, y = load_data()
julia> art = SFAM()
SFAM
    opts: opts_SFAM
    ...
julia> train!(art, x, y)