Developer Index

This page lists the types and functions that are internal to the AdaptiveResonance.jl package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

AdaptiveResonance.activation

AdaptiveResonance.activation

AdaptiveResonance.activation_match!

AdaptiveResonance.activation_match!

AdaptiveResonance.art_match

AdaptiveResonance.art_match

AdaptiveResonance.competition_kernel

AdaptiveResonance.create_category

AdaptiveResonance.create_category

AdaptiveResonance.ddt_x

AdaptiveResonance.ddt_y

AdaptiveResonance.ddt_z

AdaptiveResonance.element_min

AdaptiveResonance.get_iterator

AdaptiveResonance.get_n_weights

AdaptiveResonance.get_n_weights_vec

AdaptiveResonance.get_sample

AdaptiveResonance.init_classify!

AdaptiveResonance.init_train!

AdaptiveResonance.init_train!

AdaptiveResonance.initialize!

AdaptiveResonance.learn

AdaptiveResonance.learn

AdaptiveResonance.learn

AdaptiveResonance.learn

AdaptiveResonance.learn!

AdaptiveResonance.learn!

AdaptiveResonance.oriented_kernel

AdaptiveResonance.set_threshold!

AdaptiveResonance.similarity

AdaptiveResonance.stopping_conditions

AdaptiveResonance.stopping_conditions

AdaptiveResonance.stopping_conditions

AdaptiveResonance.stopping_conditions

AdaptiveResonance.stopping_conditions

AdaptiveResonance.surround_kernel

AdaptiveResonance.update_iter

Types

AdaptiveResonance.ARTIterator

Constants

Docs

Documentation for all internal names are listed below.

AdaptiveResonance.ARTIterator — Type

ARTIterator

Description

const ARTIterator = Union{UnitRange, ProgressBar}

Acceptable iterators for ART module training and inference

AdaptiveResonance.activation — Method

activation(
    art::DAM,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Default ARTMAP's choice-by-difference activation function.

Method List / Definition Locations

activation(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/DAM.jl:265.

AdaptiveResonance.activation — Method

activation(
    art::SFAM,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Returns the activation value of the Simple Fuzzy ARTMAP module with weight W and sample x.

Method List / Definition Locations

activation(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:281.

AdaptiveResonance.activation_match! — Method

activation_match!(
    art::DVFA,
    x::AbstractVector{T} where T<:Real
)

Summary

Compute and store the activation and match values for the DVFA module.

Method List / Definition Locations

activation_match!(art, x)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:325.

AdaptiveResonance.activation_match! — Method

activation_match!(
    art::FuzzyART,
    x::AbstractVector{T} where T<:Real
)

Summary

Computes the activation and match functions of the art module against sample x.

Arguments

art::FuzzyART: the FuzzyART module to compute the activation and match values for all weights.

x::RealVector: the sample to compute the activation and match functions against.

Examples

julia> my_FuzzyART = FuzzyART()
FuzzyART
    opts: opts_FuzzyART
    ...
julia> x, y = load_data()
julia> train!(my_FuzzyART, x)
julia> x_sample = x[:, 1]
julia> activation_match!(my_FuzzyART, x_sample)

Method List / Definition Locations

activation_match!(art, x)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:390.

AdaptiveResonance.art_match — Method

art_match(
    art::DAM,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Returns the match function for the Default ARTMAP module with weight W and sample x.

Method List / Definition Locations

art_match(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/DAM.jl:284.

AdaptiveResonance.art_match — Method

art_match(
    art::SFAM,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Returns the match function for the Simple Fuzzy ARTMAP module with weight W and sample x.

Method List / Definition Locations

art_match(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:290.

AdaptiveResonance.competition_kernel — Method

competition_kernel(l::Integer, k::Integer; sign) -> Any

Summary

Competition kernel for ARTSCENE: Stage 5.

Method List / Definition Locations

competition_kernel(l, k; sign)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:198.

AdaptiveResonance.create_category — Method

create_category(
    art::DDVFA,
    sample::AbstractVector{T} where T<:Real,
    label::Integer
) -> Vector{FuzzyART}

Summary

Create a new category by appending and initializing a new FuzzyART node to F2.

Arguments

art::DDVFA: the DDVFA module to create a new FuzzyART category in.

sample::RealVector: the sample to use for instantiating the new category.

label::Integer: the new label to use for the new category.

Method List / Definition Locations

create_category(art, sample, label)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:359.

AdaptiveResonance.create_category — Method

create_category(
    art::FuzzyART,
    x::AbstractVector{T} where T<:Real,
    y::Integer
) -> Vector{Int64}

Summary

Creates a category for the FuzzyART module, expanding the weights and incrementing the category labels.

Arguments

art::FuzzyART: the FuzzyART module to add a category to.

x::RealVector: the sample to use for adding a category.

y::Integer: the new label for the new category.

Method List / Definition Locations

create_category(art, x, y)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:360.

AdaptiveResonance.ddt_x — Method

ddt_x(
    x::AbstractArray{T} where T<:Real,
    image::AbstractArray{T} where T<:Real,
    sigma_s::AbstractArray{T} where T<:Real,
    distributed::Bool
) -> Any

Summary

Time rate of change of LGN network (ARTSCENE Stage 2).

Method List / Definition Locations

ddt_x(x, image, sigma_s, distributed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:38.

AdaptiveResonance.ddt_y — Method

ddt_y(
    y::AbstractArray{T} where T<:Real,
    X_plus::AbstractArray{T} where T<:Real,
    X_minus::AbstractArray{T} where T<:Real,
    alpha::Real,
    distributed::Bool
) -> Any

Summary

Shunting equation for ARTSCENE Stage 3.

Method List / Definition Locations

ddt_y(y, X_plus, X_minus, alpha, distributed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:114.

AdaptiveResonance.ddt_z — Method

ddt_z(z::AbstractArray{T} where T<:Real; distributed) -> Any

Summary

Time rate of change for ARTSCENE: Stage 5.

Method List / Definition Locations

ddt_z(z; distributed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:214.

AdaptiveResonance.element_min — Method

element_min(
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Returns the element-wise minimum between sample x and weight W.

Arguments

x::RealVector: the input sample.

W::RealVector: the weight vector to compare the sample against.

Method List / Definition Locations

element_min(x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:200.

AdaptiveResonance.get_iterator — Method

get_iterator(opts::ARTOpts, n_samples::Integer) -> Any

Summary

Creates an iterator object according to the ART/ARTMAP modules display settings for batch iteration.

Arguments

opts::ARTOpts: the ART/ARTMAP module's options containing display settings.

n_samples::Integer: the number of iterations to create the iterator for.

Method List / Definition Locations

get_iterator(opts, n_samples)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:419.

AdaptiveResonance.get_n_weights — Method

get_n_weights(art::DDVFA) -> Int64

Summary

Convenience function; return the sum total number of weights in the DDVFA module.

Method List / Definition Locations

get_n_weights(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:475.

AdaptiveResonance.get_n_weights_vec — Method

get_n_weights_vec(art::DDVFA) -> Vector{Int64}

Summary

Convenience function; return the number of weights in each category as a vector.

Arguments

art::DDVFA: the DDVFA module to get all of the weights from as a list.

Method List / Definition Locations

get_n_weights_vec(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:468.

AdaptiveResonance.get_sample — Method

get_sample(
    x::AbstractMatrix{T} where T<:Real,
    i::Integer
) -> Any

Summary

Returns a sample from data array x at sample location i. This function implements the convention that columns are samples while rows are features within samples.

Arguments

x::RealMatrix: the batch of data to grab a sample from.

i::Integer: the index to get the sample from.

Method List / Definition Locations

get_sample(x, i)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:454.

AdaptiveResonance.init_classify! — Method

init_classify!(
    x::AbstractArray{T} where T<:Real,
    art::ARTModule,
    preprocessed::Bool
) -> Any

Summary

Initializes the classification loop for batch inference.

Arguments

x::RealArray: the data that is used for inference.

art::ARTModule: the ART/ARTMAP module that will be used for inference.

preprocessed::Bool: required flag for if the data has already been complement coded and normalized.

Method List / Definition Locations

init_classify!(x, art, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:519.

AdaptiveResonance.init_train! — Method

init_train!(
    x::AbstractMatrix{T} where T<:Real,
    art::ARTModule,
    preprocessed::Bool
) -> Any

Summary

Initializes the training loop for batch learning.

Arguments

x::RealMatrix: the data that is used for training.

art::ARTModule: the ART/ARTMAP that will be trained.

preprocessed::Bool: required flag for if the data has already been complement coded and normalized.

Method List / Definition Locations

init_train!(x, art, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:501.

AdaptiveResonance.init_train! — Method

init_train!(
    x::AbstractVector{T} where T<:Real,
    art::ARTModule,
    preprocessed::Bool
) -> AbstractVector{T} where T<:Real

Summary

Initializes the module for training in a single iteration.

The purpose of this function is mainly to handle the conditions of complement coding. Fails if the module was incorrectly set up or if the module was not setup and the data was not preprocessed.

Arguments

x::RealVector: the sample used for initialization.

art::ARTModule: the ART/ARTMAP module that will be trained on the sample.

preprocessed::Bool: a required flag for if the sample has already been complement coded and normalized.

Method List / Definition Locations

init_train!(x, art, preprocessed)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:470.

AdaptiveResonance.initialize! — Method

initialize!(
    art::FuzzyART,
    x::AbstractVector{T} where T<:Real;
    y
) -> Vector{Int64}

Summary

Initializes a FuzzyART learner with an intial sample 'x'.

This function is used during the first training iteration when the FuzzyART module is empty.

Arguments

art::FuzzyART: the FuzzyART module to initialize.

x::RealVector: the sample to use for initialization.

y::Integer=0: the optional new label for the first weight of the FuzzyART module. If not specified, defaults the new label to 1.

Examples

julia> my_FuzzyART = FuzzyART()
FuzzyART
    opts: opts_FuzzyART
    ...
julia> initialize!(my_FuzzyART, [1 2 3 4])

Method List / Definition Locations

initialize!(art, x; y)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:240.

AdaptiveResonance.learn! — Method

learn!(
    art::DVFA,
    x::AbstractVector{T} where T<:Real,
    index::Integer
) -> Any

Summary

In place learning function.

Method List / Definition Locations

learn!(art, x, index)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:346.

AdaptiveResonance.learn! — Method

learn!(
    art::FuzzyART,
    x::AbstractVector{T} where T<:Real,
    index::Integer
) -> Int64

Summary

In place learning function with instance counting.

Arguments

art::FuzzyART: the FuzzyART module to update.

x::RealVector: the sample to learn from.

index::Integer: the index of the FuzzyART weight to update.

Method List / Definition Locations

learn!(art, x, index)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:426.

AdaptiveResonance.learn — Method

learn(
    art::DAM,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Returns a single updated weight for the Default ARTMAP module for weight vector W and sample x.

Method List / Definition Locations

learn(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/DAM.jl:276.

AdaptiveResonance.learn — Method

learn(
    art::DVFA,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Return the modified weight of the DVFA module conditioned by sample x.

Method List / Definition Locations

learn(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:338.

AdaptiveResonance.learn — Method

learn(
    art::FuzzyART,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Return the modified weight of the art module conditioned by sample x.

Arguments

art::FuzzyART: the FuzzyART module containing learning options.

x::RealVector: the sample to learn from.

W::RealVector: the weight vector to update against the sample.

Method List / Definition Locations

learn(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:413.

AdaptiveResonance.learn — Method

learn(
    art::SFAM,
    x::AbstractVector{T} where T<:Real,
    W::AbstractVector{T} where T<:Real
) -> Any

Summary

Returns a single updated weight for the Simple Fuzzy ARTMAP module for weight vector W and sample x.

Method List / Definition Locations

learn(art, x, W)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:272.

AdaptiveResonance.oriented_kernel — Method

oriented_kernel(
    i::Integer,
    j::Integer,
    p::Integer,
    q::Integer,
    k::Integer,
    sigma_h::Real,
    sigma_v::Real;
    sign
) -> Any

Summary

Oriented, elongated, spatially offset kernel G for ARTSCENE Stage 3.

Method List / Definition Locations

oriented_kernel(i, j, p, q, k, sigma_h, sigma_v; sign)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:92.

AdaptiveResonance.set_threshold! — Method

Summary

Sets the match threshold of the ART/ARTMAP module as a function of the vigilance parameter.

Depending on selected ART/ARTMAP module and its options, this may be a function of other parameters as well.

Arguments

art::ARTModule: the ART/ARTMAP module for setting a new threshold.

Method List / Definition Locations

set_threshold!(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:219.

set_threshold!(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:192.

set_threshold!(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:213.

AdaptiveResonance.similarity — Method

similarity(
    method::AbstractString,
    F2::FuzzyART,
    field_name::AbstractString,
    sample::AbstractVector{T} where T<:Real
) -> Any

Summary

Compute the similarity metric depending on method with explicit comparisons for the field name.

Arguments

method::AbstractString: the selected DDVFA linkage method.

F2::FuzzyART: the FuzzyART module to compute the linkage method within.

field_name::AbstractString: the activation or match value to compute, field_name ∈ ["T", "M"]

sample::RealVector: the sample to use for computing the linkage to the F2 module, sample ∈ DDVFA_METHODS.

Method List / Definition Locations

similarity(method, F2, field_name, sample)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:389.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::DAM) -> Bool

Summary

Stopping conditions for Default ARTMAP, checked at the end of every epoch.

Method List / Definition Locations

stopping_conditions(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/DAM.jl:257.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::DDVFA) -> Bool

Summary

Stopping conditions for Distributed Dual Vigilance Fuzzy ARTMAP.

Returns true if there is no change in weights during the epoch or the maxmimum epochs has been reached.

Arguments

art::DDVFA: the DDVFA module for checking stopping conditions.

Method List / Definition Locations

stopping_conditions(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DDVFA.jl:375.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::DVFA) -> Bool

Summary

Stopping conditions for a DVFA module.

Method List / Definition Locations

stopping_conditions(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/DVFA.jl:354.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::FuzzyART) -> Bool

Summary

Stopping conditions for a FuzzyART module.

Arguments

art::FuzzyART: the FuzzyART module to check stopping conditions for.

Method List / Definition Locations

stopping_conditions(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ART/FuzzyART.jl:438.

AdaptiveResonance.stopping_conditions — Method

stopping_conditions(art::SFAM) -> Bool

Summary

Stopping conditions for Simple Fuzzy ARTMAP, checked at the end of every epoch.

Method List / Definition Locations

stopping_conditions(art)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/SFAM.jl:263.

AdaptiveResonance.surround_kernel — Method

surround_kernel(
    i::Integer,
    j::Integer,
    p::Integer,
    q::Integer,
    scale::Integer
) -> Any

Summary

Surround kernel S function for ARTSCENE Stage 2.

Method List / Definition Locations

surround_kernel(i, j, p, q, scale)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/ARTMAP/ARTSCENE.jl:31.

AdaptiveResonance.update_iter — Method

update_iter(
    art::ARTModule,
    iter::Union{ProgressBars.ProgressBar, UnitRange},
    i::Integer
) -> Union{Nothing, String}

Summary

Updates the iteration of the ART/ARTMAP module, training or inference, according to its display settings.

Arguments

art::ARTModule: the ART/ARTMAP module being iterated upon.

iter::ARTIterator: the iterator object used in the training/inference loop.

i::Integer: the iteration during training/inference that the iterator should be updated to.

Method List / Definition Locations

update_iter(art, iter, i)

defined at /home/runner/work/AdaptiveResonance.jl/AdaptiveResonance.jl/src/common.jl:437.