Index

This page lists the core methods and types of the AdaptiveResonance.jl package. The Methods section lists the public methods for the package that use the modules in Types. Each of these entries link to the docstrings in the Docs section.

ART modules document their internal working parameters and references, while their hyperparameters/options are documented under their corresponding option structs opts_....

Methods

AdaptiveResonance.classify

AdaptiveResonance.complement_code

AdaptiveResonance.data_setup!

AdaptiveResonance.get_data_characteristics

AdaptiveResonance.get_data_shape

AdaptiveResonance.get_n_samples

AdaptiveResonance.linear_normalization

AdaptiveResonance.performance

AdaptiveResonance.train!

Types

AdaptiveResonance.ART

AdaptiveResonance.ARTMAP

AdaptiveResonance.ARTModule

AdaptiveResonance.ARTOpts

AdaptiveResonance.DAM

AdaptiveResonance.DDVFA

AdaptiveResonance.DVFA

AdaptiveResonance.DataConfig

AdaptiveResonance.FAM

AdaptiveResonance.FuzzyART

AdaptiveResonance.SFAM

AdaptiveResonance.opts_DAM

AdaptiveResonance.opts_DDVFA

AdaptiveResonance.opts_DVFA

AdaptiveResonance.opts_FAM

AdaptiveResonance.opts_FuzzyART

AdaptiveResonance.opts_SFAM

Docs

AdaptiveResonance — Module

Main module for AdaptiveResonance.jl, a Julia package of adaptive resonance theory algorithms.

This module exports all of the ART modules, options, and utilities used by the AdaptiveResonance.jl package.

Exports

ART

ARTMAP

ARTModule

ARTOpts

DAM

DDVFA

DVFA

DataConfig

FAM

FuzzyART

SFAM

artscene_filter

classify

color_to_gray

complement_code

contrast_insensitive_oriented_filtering

contrast_normalization

contrast_sensitive_oriented_filtering

data_setup!

get_W

get_data_characteristics

get_data_shape

get_n_samples

linear_normalization

opts_DAM

opts_DDVFA

opts_DVFA

opts_FAM

opts_FuzzyART

opts_SFAM

orientation_competition

patch_orientation_color

performance

train!

AdaptiveResonance.train! — Function

train!(art::ARTMAP, x::RealMatrix, y::IntegerVector, preprocessed::Bool=false)

Train the ARTMAP model on a batch of data 'x' with supervisory labels 'y.'

Arguments

art::ARTMAP: the supervised ARTMAP model to train.

x::RealMatrix: the 2-D dataset containing columns of samples with rows of features.

y::IntegerVector: labels for supervisory training.

preprocessed::Bool=false: flag, if the data has already been complement coded or not.

train!(art::ARTMAP, x::RealVector, y::Integer ; preprocessed::Bool=false)

Train the supervised ARTMAP model on a single sample of features 'x' with supervisory label 'y'.

Arguments

art::ARTMAP: the supervised ART model to train.

x::RealVector: the single sample feature vector to train upon.

y::Integer: the label for supervisory training.

preprocessed::Bool=false: optional, flag if the data has already been complement coded or not.

train!(art::ART, x::RealMatrix ; y::IntegerVector=Vector{Int}(), preprocessed::Bool=false)

Train the ART model on a batch of data 'x' with optional supervisory labels 'y.'

Arguments

art::ART: the unsupervised ART model to train.

x::RealMatrix: the 2-D dataset containing columns of samples with rows of features.

y::IntegerVector=Vector{Int}(): optional, labels for simple supervisory training.

preprocessed::Bool=false: optional, flag if the data has already been complement coded or not.

train!(art::ART, x::RealVector ; y::Integer=0, preprocessed::Bool=false)

Train the ART model on a single sample of features 'x' with an optional supervisory label.

Arguments

art::ART: the unsupervised ART model to train.

x::RealVector: the single sample feature vector to train upon.

y::Integer=0: optional, a label for simple supervisory training.

preprocessed::Bool=false: optional, flag if the data has already been complement coded or not.

AdaptiveResonance.classify — Function

classify(art::ARTModule, x::RealMatrix ; preprocessed::Bool=false, get_bmu::Bool=false)

Predict categories of 'x' using the ART model.

Returns predicted categories 'y_hat.'

Arguments

art::ARTModule: ART or ARTMAP module to use for batch inference.

x::RealMatrix: the 2-D dataset containing columns of samples with rows of features.

preprocessed::Bool=false: flag, if the data has already been complement coded or not.

get_bmu::Bool=false, flag, if the model should return the best-matching-unit label in the case of total mismatch.

Examples

julia> my_DDVFA = DDVFA()
DDVFA
    opts: opts_DDVFA
    ...
julia> x, y = load_data()
julia> train!(my_DDVFA, x)
julia> y_hat = classify(my_DDVFA, y)

classify(art::ARTModule, x::RealVector ; preprocessed::Bool=false, get_bmu::Bool=false)

Predict categories of a single sample of features 'x' using the ART model.

Returns predicted category 'y_hat.'

Arguments

art::ARTModule: ART or ARTMAP module to use for batch inference.

x::RealVector: the single sample of features to classify.

preprocessed::Bool=false: optional, flag if the data has already been complement coded or not.

get_bmu::Bool=false: optional, flag if the model should return the best-matching-unit label in the case of total mismatch.

AdaptiveResonance.data_setup! — Function

data_setup!(config::DataConfig, data::RealMatrix)

Sets up the data config for the ART module before training.

data_setup!(art::ARTModule, data::RealMatrix)

Convenience method for setting up the DataConfig of an ART module in advance.

AdaptiveResonance.performance — Function

performance(y_hat::IntegerVector, y::IntegerVector)

Returns the categorization performance of y_hat against y.

AdaptiveResonance.complement_code — Function

complement_code(data::RealArray ; config::DataConfig=DataConfig())

Normalize the data x to [0, 1] and returns the augmented vector [x, 1 - x].

AdaptiveResonance.get_data_characteristics — Function

get_data_characteristics(data::RealArray ; config::DataConfig=DataConfig())

Get the characteristics of the data, taking account if a data config is passed.

If no DataConfig is passed, then the data characteristics come from the array itself. Otherwise, use the config for the statistics of the data and the data array for the number of samples.

AdaptiveResonance.linear_normalization — Function

linear_normalization(data::RealVector ; config::DataConfig=DataConfig())

Normalize the data to the range [0, 1] along each feature.

linear_normalization(data::RealMatrix ; config::DataConfig=DataConfig())

Normalize the data to the range [0, 1] along each feature.

AdaptiveResonance.get_data_shape — Function

get_data_shape(data::RealArray)

Returns the correct feature dimension and number of samples.

AdaptiveResonance.get_n_samples — Function

get_n_samples(data::RealArray)

Returns the number of samples, accounting for 1-D and 2-D arrays.

AdaptiveResonance.DDVFA — Type

DDVFA <: ART

Distributed Dual Vigilance Fuzzy ARTMAP module struct.

For module options, see AdaptiveResonance.opts_DDVFA.

Option Parameters

opts::opts_DDVFA: DDVFA options struct.

subopts::opts_FuzzyART: FuzzyART options struct used for all F2 nodes.

config::DataConfig: data configuration struct.

Working Parameters

threshold::Float: operating module threshold value, a function of the vigilance parameter.

F2::Vector{FuzzyART}: list of F2 nodes (themselves FuzzyART modules).

labels::IntegerVector: incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

n_categories::Int: number of total categories.

epoch::Int: current training epoch.

T::Float: winning activation value from most recent sample.

M::Float: winning match value from most recent sample.

References

L. E. Brito da Silva, I. Elnabarawy, and D. C. Wunsch, “Distributed dual vigilance fuzzy adaptive resonance theory learns online, retrieves arbitrarily-shaped clusters, and mitigates order dependence,” Neural Networks, vol. 121, pp. 208-228, 2020, doi: 10.1016/j.neunet.2019.08.033.

G. Carpenter, S. Grossberg, and D. Rosen, "Fuzzy ART: Fast stable learning and categorization of analog patterns by an adaptive resonance system," Neural Networks, vol. 4, no. 6, pp. 759-771, 1991.

AdaptiveResonance.DVFA — Type

DVFA <: ART

Dual Vigilance Fuzzy ARTMAP module struct.

For module options, see AdaptiveResonance.opts_DVFA.

Option Parameters

opts::opts_DVFA: DVFA options struct.

config::DataConfig: data configuration struct.

Working Parameters

threshold_ub::Float: operating upper bound module threshold value, a function of the upper bound vigilance parameter.

threshold_lb::Float: operating lower bound module threshold value, a function of the lower bound vigilance parameter.

labels::IntegerVector: incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

W::RealMatrix: category weight matrix.

T::RealVector: activation values for every weight for a given sample.

M::RealVector: match values for every weight for a given sample.

n_categories::Int: number of category weights (F2 nodes).

n_clusters::Int: number of labeled clusters, may be lower than n_categories

epoch::Int: current training epoch.

References:

L. E. Brito da Silva, I. Elnabarawy and D. C. Wunsch II, "Dual Vigilance Fuzzy ART," Neural Networks Letters. To appear.

G. Carpenter, S. Grossberg, and D. Rosen, "Fuzzy ART: Fast stable learning and categorization of analog patterns by an adaptive resonance system," Neural Networks, vol. 4, no. 6, pp. 759-771, 1991.

AdaptiveResonance.FuzzyART — Type

FuzzyART <: ART

Gamma-Normalized Fuzzy ART learner struct

For module options, see AdaptiveResonance.opts_FuzzyART.

Option Parameters

opts::opts_FuzzyART: FuzzyART options struct.

config::DataConfig: data configuration struct.

Working Parameters

threshold::Float: operating module threshold value, a function of the vigilance parameter.

labels::IntegerVector: incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

T::RealVector: activation values for every weight for a given sample.

M::RealVector: match values for every weight for a given sample.

W::RealMatrix: category weight matrix.

n_instance::IntegerVector: number of weights associated with each category.

n_categories::Int: number of category weights (F2 nodes).

epoch::Int: current training epoch.

References

G. Carpenter, S. Grossberg, and D. Rosen, "Fuzzy ART: Fast stable learning and categorization of analog patterns by an adaptive resonance system," Neural Networks, vol. 4, no. 6, pp. 759-771, 1991.

AdaptiveResonance.DAM — Type

DAM <: ARTMAP

Default ARTMAP struct.

For module options, see AdaptiveResonance.opts_DAM.

Option Parameters

opts::opts_DAM: Default ARTMAP options struct.

config::DataConfig: data configuration struct.

Working Parameters

W::RealMatrix: category weight matrix.

labels::IntegerVector: incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

n_categories::Int: number of category weights (F2 nodes).

epoch::Int: current training epoch.

References

G. P. Amis and G. A. Carpenter, “Default ARTMAP 2,” IEEE Int. Conf. Neural Networks - Conf. Proc., vol. 2, no. September 2007, pp. 777-782, Mar. 2007, doi: 10.1109/IJCNN.2007.4371056.

AdaptiveResonance.FAM — Type

FAM <: ARTMAP

Fuzzy ARTMAP struct.

For module options, see AdaptiveResonance.opts_FAM.

Option Parameters

opts::opts_FAM: Fuzzy ARTMAP options struct.

config::DataConfig: data configuration struct.

Working Parameters

W::RealMatrix: category weight matrix.

labels::IntegerVector: incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

n_categories::Int: number of category weights (F2 nodes).

epoch::Int: current training epoch.

References

G. A. Carpenter, S. Grossberg, N. Markuzon, J. H. Reynolds, and D. B. Rosen, “Fuzzy ARTMAP: A Neural Network Architecture for Incremental Supervised Learning of Analog Multidimensional Maps,” IEEE Trans. Neural Networks, vol. 3, no. 5, pp. 698-713, 1992, doi: 10.1109/72.159059.

AdaptiveResonance.SFAM — Type

SFAM <: ARTMAP

Simple Fuzzy ARTMAP struct.

For module options, see AdaptiveResonance.opts_SFAM.

Option Parameters

opts::opts_SFAM: Simplified Fuzzy ARTMAP options struct.

config::DataConfig: data configuration struct.

Working Parameters

W::RealMatrix: category weight matrix.

labels::IntegerVector: incremental list of labels corresponding to each F2 node, self-prescribed or supervised.

n_categories::Int: number of category weights (F2 nodes).

epoch::Int: current training epoch.

References

G. A. Carpenter, S. Grossberg, N. Markuzon, J. H. Reynolds, and D. B. Rosen, “Fuzzy ARTMAP: A Neural Network Architecture for Incremental Supervised Learning of Analog Multidimensional Maps,” IEEE Trans. Neural Networks, vol. 3, no. 5, pp. 698-713, 1992, doi: 10.1109/72.159059.

AdaptiveResonance.opts_DDVFA — Type

opts_DDVFA(;kwargs)

Distributed Dual Vigilance Fuzzy ART options struct.

Keyword Arguments

rho_lb::Float: lower-bound vigilance value, [0, 1], default 0.7.

rho_ub::Float: upper-bound vigilance value, [0, 1], default 0.85.

alpha::Float: choice parameter, alpha > 0, default 1e-3.

beta::Float: learning parameter, (0, 1], default 1.0.

gamma::Float: "pseudo" kernel width, gamma >= 1, default 3.0.

gamma_ref::Float: "reference" kernel width, 0 <= gamma_ref < gamma, default 1.0.

method::String: similarity method (activation and match):

single, average, complete, median, weighted, or centroid, default single.

display::Bool: display flag, default true.

max_epoch::Int: maximum number of epochs during training, default 1.

gamma_normalization::Bool: normalize the threshold by the feature dimension, default true.

AdaptiveResonance.opts_DVFA — Type

opts_DVFA(;kwargs)

Dual Vigilance Fuzzy ART options struct.

Keyword Arguments

rho_lb::Float: lower-bound vigilance value, [0, 1], default 0.55.

rho_ub::Float: upper-bound vigilance value, [0, 1], default 0.75.

alpha::Float: choice parameter, alpha > 0, default 1e-3.

beta::Float: learning parameter, (0, 1], default 1.0.

display::Bool: display flag, default true.

max_epoch::Int: maximum number of epochs during training, default 1.

AdaptiveResonance.opts_FuzzyART — Type

opts_FuzzyART(;kwargs)

Gamma-Normalized Fuzzy ART options struct.

Keyword Arguments

rho::Float: vigilance value, [0, 1], default 0.6.

alpha::Float: choice parameter, alpha > 0, default 1e-3.

beta::Float: learning parameter, (0, 1], default 1.0.

gamma::Float: "pseudo" kernel width, gamma >= 1, default 3.0.

gamma_ref::Float: "reference" kernel width, 0 <= gamma_ref < gamma, default 1.0.

display::Bool: display flag, default true.

max_epoch::Int: maximum number of epochs during training, default 1.

gamma_normalization::Bool: normalize the threshold by the feature dimension, default false.

AdaptiveResonance.opts_DAM — Type

opts_DAM(;kwargs)

Implements a Default ARTMAP learner's options.

Keyword Arguments

rho::Float: vigilance value, [0, 1], default 0.75.

alpha::Float: choice parameter, alpha > 0, default 1e-7.

epsilon::Float: match tracking parameter, (0, 1), default 1e-3

beta::Float: learning parameter, (0, 1], default 1.0.

uncommitted::Bool: uncommitted node flag, default true.

display::Bool: display flag, default true.

max_epoch::Int: maximum number of epochs during training, default 1.

AdaptiveResonance.opts_FAM — Type

opts_FAM(;kwargs)

Implements a Fuzzy ARTMAP learner's options.

Keyword Arguments

rho::Float: vigilance value, [0, 1], default 0.6.

alpha::Float: choice parameter, alpha > 0, default 1e-7.

epsilon::Float: match tracking parameter, (0, 1), default 1e-3

beta::Float: learning parameter, (0, 1], default 1.0.

uncommitted::Bool: uncommitted node flag, default true.

display::Bool: display flag, default true.

max_epoch::Int: maximum number of epochs during training, default 1.

AdaptiveResonance.opts_SFAM — Type

opts_SFAM(;kwargs)

Implements a Simple Fuzzy ARTMAP learner's options.

Keyword Arguments

rho::Float: vigilance value, [0, 1], default 0.75.

alpha::Float: choice parameter, alpha > 0, default 1e-7.

epsilon::Float: match tracking parameter, (0, 1), default 1e-3

beta::Float: learning parameter, (0, 1], default 1.0.

uncommitted::Bool: uncommitted node flag, default true.

display::Bool: display flag, default true.

max_epoch::Int: maximum number of epochs during training, default 1.

AdaptiveResonance.DataConfig — Type

DataConfig

Container to standardize training/testing data configuration.

Parameters

setup::Bool: flag if data has been setup yet or not.

mins::RealVector: list of minimum values for each feature.

maxs::RealVector: list of maximum values for each feature.

dim::Int: dimensionality of the feature vectors (i.e., number of features).

dim_comp::Int complement coded feature dimensionality, twice the size of dim.

AdaptiveResonance.ARTModule — Type

ARTModule

Abstract supertype for both ART (unsupervised) and ARTMAP (supervised) modules.

AdaptiveResonance.ART — Type

ART <: ARTModule

Abstract supertype for all default unsupervised ART modules.

AdaptiveResonance.ARTMAP — Type

ARTMAP <: ARTModule

Abstract supertype for all supervised ARTMAP modules.

AdaptiveResonance.ARTOpts — Type

ARTOpts

Abstract supertype for all ART module options.