Index

AdaptiveResonance.DAM — Type

DAM <: ARTMAP

Default ARTMAP struct.

AdaptiveResonance.DAM — Method

DAM(opts)

Implements a Default ARTMAP learner with specified options.

Examples

julia> opts = opts_DAM()
julia> DAM(opts)
DAM
    opts: opts_DAM
    ...

AdaptiveResonance.DAM — Method

DAM(;kwargs...)

Implements a Default ARTMAP learner with keyword arguments.

Examples

julia> DAM()
DAM
    opts: opts_DAM
    ...

AdaptiveResonance.DDVFA — Type

DDVFA <: ART

Distributed Dual Vigilance Fuzzy ARTMAP module struct.

Examples

julia> DDVFA()
DDVFA
    opts: opts_DDVFA
    subopts::opts_FuzzyART
    ...

AdaptiveResonance.DDVFA — Method

DDVFA(opts::opts_DDVFA)

Implements a DDVFA learner with specified options.

Examples

julia> my_opts = opts_DDVFA()
julia> DDVFA(my_opts)
DDVFA
    opts: opts_DDVFA
    subopts: opts_FuzzyART
    ...

AdaptiveResonance.DDVFA — Method

DDVFA(;kwargs...)

Implements a DDVFA learner with keyword arguments.

Examples

julia> DDVFA(rho_lb=0.4, rho_ub = 0.75)
DDVFA
    opts: opts_DDVFA
    subopts: opts_FuzzyART
    ...

AdaptiveResonance.DVFA — Type

DVFA <: ART

Dual Vigilance Fuzzy ARTMAP module struct.

Examples

julia> DVFA()
DVFA
    opts: opts_DVFA
    ...

AdaptiveResonance.DVFA — Method

DVFA(opts::opts_DVFA)

Implements a DVFA learner with specified options.

Examples

julia> my_opts = opts_DVFA()
julia> DVFA(my_opts)
DVFA
    opts: opts_DVFA
    ...

AdaptiveResonance.DVFA — Method

DVFA(;kwargs...)

Implements a DVFA learner with keyword arguments.

Examples

julia> DVFA(rho=0.7)
DVFA
    opts: opts_DDVFA
    ...

AdaptiveResonance.DataConfig — Type

DataConfig

Conatiner to standardize training/testing data configuration.

AdaptiveResonance.DataConfig — Method

DataConfig(mins::RealVector, maxs::RealVector)

Convenience constructor for DataConfig, requiring only mins and maxs of the features.

This constructor is used when the mins and maxs differ across features. The dimension is inferred by the length of the mins and maxs.

AdaptiveResonance.DataConfig — Method

DataConfig(data::RealMatrix)

Convenience constructor for DataConfig, requiring only the data matrix.

AdaptiveResonance.DataConfig — Method

DataConfig(min::Real, max::Real, dim::Int)

Convenience constructor for DataConfig, requiring only a global min, max, and dim.

This constructor is used in the case that the feature mins and maxs are all the same respectively.

AdaptiveResonance.DataConfig — Method

DataConfig()

Default constructor for a data configuration, not set up.

AdaptiveResonance.FAM — Type

FAM <: ARTMAP

Fuzzy ARTMAP struct.

AdaptiveResonance.FAM — Method

FAM(opts)

Implements a Fuzzy ARTMAP learner with specified options.

Examples

julia> opts = opts_FAM()
julia> FAM(opts)
FAM
    opts: opts_FAM
    ...

AdaptiveResonance.FAM — Method

FAM(;kwargs...)

Implements a Fuzzy ARTMAP learner with keyword arguments.

Examples

julia> FAM(rho=0.7)
FAM
    opts: opts_FAM
    ...

AdaptiveResonance.FuzzyART — Type

FuzzyART <: ART

Gamma-Normalized Fuzzy ART learner struct

Examples

julia> FuzzyART()
FuzzyART
    opts: opts_FuzzyART
    ...

AdaptiveResonance.FuzzyART — Method

FuzzyART(opts::opts_FuzzyART, sample::RealVector)

Create and initialize a FuzzyART with a single sample in one step.

AdaptiveResonance.FuzzyART — Method

FuzzyART(opts::opts_FuzzyART)

Implements a Gamma-Normalized Fuzzy ART learner with specified options.

Examples

julia> FuzzyART(opts)
FuzzyART
    opts: opts_FuzzyART
    ...

AdaptiveResonance.FuzzyART — Method

FuzzyART(;kwargs...)

Implements a Gamma-Normalized Fuzzy ART learner with keyword arguments.

Examples

julia> FuzzyART(rho=0.7)
FuzzyART
    opts: opts_FuzzyART
    ...

AdaptiveResonance.SFAM — Type

SFAM <: ARTMAP

Simple Fuzzy ARTMAP struct.

AdaptiveResonance.SFAM — Method

SFAM(opts)

Implements a Simple Fuzzy ARTMAP learner with specified options.

Examples

julia> opts = opts_SFAM()
julia> SFAM(opts)
SFAM
    opts: opts_SFAM
    ...

AdaptiveResonance.SFAM — Method

SFAM(;kwargs...)

Implements a Simple Fuzzy ARTMAP learner with keyword arguments.

Examples

julia> SFAM()
SFAM
    opts: opts_SFAM
    ...

AdaptiveResonance.opts_DAM — Type

opts_DAM()

Implements a Default ARTMAP learner's options.

Examples

julia> my_opts = opts_DAM()

AdaptiveResonance.opts_DDVFA — Type

opts_DDVFA()

Distributed Dual Vigilance Fuzzy ART options struct.

Examples

julia> my_opts = opts_DDVFA()

AdaptiveResonance.opts_DVFA — Type

opts_DVFA()

Dual Vigilance Fuzzy ART options struct.

Examples

julia> my_opts = opts_DVFA()

AdaptiveResonance.opts_FAM — Type

opts_FAM()

Implements a Fuzzy ARTMAP learner's options.

Examples

julia> my_opts = opts_FAM()

AdaptiveResonance.opts_FuzzyART — Type

opts_FuzzyART()

Gamma-Normalized Fuzzy ART options struct.

Examples

julia> opts_FuzzyART()
Initialized FuzzyART

AdaptiveResonance.opts_SFAM — Type

opts_SFAM()

Implements a Simple Fuzzy ARTMAP learner's options.

Examples

julia> my_opts = opts_SFAM()

AdaptiveResonance.artscene_filter — Method

artscene_filter(raw_image::Array{T, 3} ;  distributed::Bool=true) where {T<:Real}

Process the full artscene filter toolchain on an image.

AdaptiveResonance.classify — Method

classify(art::ARTModule, x::RealMatrix ; preprocessed::Bool=false, get_bmu::Bool=false)

Predict categories of 'x' using the ART model.

Returns predicted categories 'y_hat.'

Arguments

art::ARTModule: ART or ARTMAP module to use for batch inference.

x::RealMatrix: the 2-D dataset containing columns of samples with rows of features.

preprocessed::Bool=false: flag, if the data has already been complement coded or not.

get_bmu::Bool=false, flag, if the model should return the best-matching-unit label in the case of total mismatch.

Examples

julia> my_DDVFA = DDVFA()
DDVFA
    opts: opts_DDVFA
    ...
julia> x, y = load_data()
julia> train!(my_DDVFA, x)
julia> y_hat = classify(my_DDVFA, y)

AdaptiveResonance.classify — Method

classify(art::DAM, x::RealArray ; preprocessed::Bool=false)

Categorize data 'x' using a trained Default ARTMAP module 'art'.

Examples

julia> x, y = load_data()
julia> x_test, y_test = load_test_data()
julia> art = DAM()
DAM
    opts: opts_DAM
    ...
julia> train!(art, x, y)
julia> classify(art, x_test)

AdaptiveResonance.classify — Method

classify(art::DDVFA, x::RealMatrix ; preprocessed::Bool=false, get_bmu::Bool=false)

Predict categories of 'x' using the DDVFA model.

Returns predicted categories 'y_hat.'

Examples

julia> my_DDVFA = DDVFA()
DDVFA
    opts: opts_DDVFA
    ...
julia> x, y = load_data()
julia> train!(my_DDVFA, x)
julia> y_hat = classify(my_DDVFA, y)

AdaptiveResonance.classify — Method

classify(art::DVFA, x::RealVector ; preprocessed::Bool=false, get_bmu::Bool=false)

Predict categories of 'x' using the DVFA model.

Returns predicted categories 'y_hat'

Examples

julia> my_DVFA = DVFA()
DVFA
    opts: opts_DVFA
    ...
julia> x, y = load_data()
julia> train!(my_DVFA, x)
julia> y_hat = classify(my_DVFA, y)

AdaptiveResonance.classify — Method

classify(art::FuzzyART, x::RealVector ; preprocessed::Bool=false, get_bmu::Bool=false)

AdaptiveResonance.classify — Method

classify(art::FuzzyART, x::RealMatrix ; preprocessed::Bool=false, get_bmu::Bool=false)

Batch predict categories of 'x' using the FuzzyART model.

Returns predicted categories 'y_hat'

Examples

julia> my_FuzzyART = FuzzyART()
FuzzyART
    opts: opts_FuzzyART
    ...
julia> x, y = load_data()
julia> train!(my_FuzzyART, x)
julia> y_hat = classify(my_FuzzyART, y)

AdaptiveResonance.classify — Method

classify(art::SFAM, x::RealArray ; preprocessed::Bool=false)

Categorize data 'x' using a trained Simple Fuzzy ARTMAP module 'art'.

Examples

julia> x, y = load_data()
julia> x_test, y_test = load_test_data()
julia> art = SFAM()
SFAM
    opts: opts_SFAM
    ...
julia> train!(art, x, y)
julia> classify(art, x_test)

AdaptiveResonance.color_to_gray — Method

color_to_gray(image::RealArray)

ARTSCENE Stage 1: Color-to-gray image transformation.

AdaptiveResonance.complement_code — Method

complement_code(data::RealArray ; config::DataConfig=DataConfig())

Normalize the data x to [0, 1] and returns the augmented vector [x, 1 - x].

AdaptiveResonance.contrast_insensitive_oriented_filtering — Method

contrast_insensitive_oriented_filtering(y::RealArray)

ARTSCENE Stage 4: Contrast-insensitive oriented filtering.

AdaptiveResonance.contrast_normalization — Method

contrast_normalization(image::RealArray ; distributed::Bool=true)

ARTSCENE Stage 2: Constrast normalization.

AdaptiveResonance.contrast_sensitive_oriented_filtering — Method

contrast_sensitive_oriented_filtering(image::RealArray, x::RealArray ; distributed::Bool=true)

ARTSCENE Stage 3: Contrast-sensitive oriented filtering.

AdaptiveResonance.data_setup! — Method

data_setup!(art::ART, data::RealMatrix)

Convenience method for setting up the DataConfig of an ART module in advance.

AdaptiveResonance.data_setup! — Method

data_setup!(config::DataConfig, data::RealMatrix)

Sets up the data config for the ART module before training.

AdaptiveResonance.get_W — Method

get_W(art::DDVFA)

Convenience functio; return a concatenated array of all DDVFA weights.

AdaptiveResonance.get_data_characteristics — Method

get_data_characteristics(data::RealArray ; config::DataConfig=DataConfig())

Get the characteristics of the data, taking account if a data config is passed.

If no DataConfig is passed, then the data characteristics come from the array itself. Otherwise, use the config for the statistics of the data and the data array for the number of samples.

AdaptiveResonance.get_data_shape — Method

get_data_shape(data::RealArray)

Returns the correct feature dimension and number of samples.

AdaptiveResonance.get_n_samples — Method

get_n_samples(data::RealArray)

Returns the number of samples, accounting for 1-D and 2-D arrays.

AdaptiveResonance.linear_normalization — Method

linear_normalization(data::RealVector ; config::DataConfig=DataConfig())

Normalize the data to the range [0, 1] along each feature.

AdaptiveResonance.linear_normalization — Method

linear_normalization(data::RealMatrix ; config::DataConfig=DataConfig())

Normalize the data to the range [0, 1] along each feature.

AdaptiveResonance.orientation_competition — Method

orientation_competition(z::RealArray)

ARTSCENE Stage 5: Orientation competition at the same position.

AdaptiveResonance.patch_orientation_color — Method

patch_orientation_color(z::RealArray, image::RealArray)

ARTSCENE Stage 6: Create patch feature vectors.

AdaptiveResonance.performance — Method

performance(y_hat::IntegerVector, y::IntegerVector)

Returns the categorization performance of y_hat against y.

AdaptiveResonance.train! — Function

train!(art::ARTMAP, x::RealMatrix, y::IntegerVector, preprocessed::Bool=false)

Train the ARTMAP model on a batch of data 'x' with supervisory labels 'y.'

Arguments

art::ARTMAP: the supervised ARTMAP model to train.

x::RealMatrix: the 2-D dataset containing columns of samples with rows of features.

y::IntegerVector: labels for supervisory training.

preprocessed::Bool=false: flag, if the data has already been complement coded or not.

AdaptiveResonance.train! — Method

train!(art::ART, x::RealMatrix ; y::IntegerVector=Vector{Int}(), preprocessed::Bool=false)

Train the ART model on a batch of data 'x' with optional supervisory labels 'y.'

Arguments

art::ART: the unsupervised ART model to train.

x::RealMatrix: the 2-D dataset containing columns of samples with rows of features.

y::IntegerVector=Vector{Int}(): optional, labels for simple supervisory training.

preprocessed::Bool=false: flag, if the data has already been complement coded or not.

AdaptiveResonance.train! — Method

train!(art::DAM, x::RealArray, y::RealArray ; preprocessed::Bool=false)

Trains a Default ARTMAP learner in a supervised manner.

Examples

julia> x, y = load_data()
julia> art = DAM()
DAM
    opts: opts_DAM
    ...
julia> train!(art, x, y)

AdaptiveResonance.train! — Method

train!(art::DDVFA, x::RealVector ; y::Integer=0, preprocessed::Bool=false)

AdaptiveResonance.train! — Method

train!(art::DVFA, x::RealVector ; y::Integer=0, preprocessed::Bool=false)

Train the DVFA module on x with optional custom category labels y.

Arguments

art::DVFA: the dual-vigilance fuzzy art module to train.

x::RealVector: the data to train on, interpreted as a single sample if x is a vector.

y::Integer=0: optional custom label to assign to the categories. If zero, ordinary incremental labels are prescribed.

AdaptiveResonance.train! — Method

train!(art::FuzzyART, x::RealVector ; y::Integer=0, preprocessed::Bool=false)

AdaptiveResonance.train! — Method

train!(art::FuzzyART, x::RealMatrix ; y::IntegerVector = Vector{Int}(), preprocessed::Bool=false)

Trains a FuzzyART learner with dataset 'x' and optional labels 'y'

Examples

julia> my_FuzzyART = FuzzyART()
FuzzyART
    opts: opts_FuzzyART
    ...
julia> x = load_data()
julia> train!(my_FuzzyART, x)

AdaptiveResonance.train! — Method

train!(art::SFAM, x::RealArray, y::RealArray ; preprocessed::Bool=false)

Trains a Simple Fuzzy ARTMAP learner in a supervised manner.

Examples

julia> x, y = load_data()
julia> art = SFAM()
SFAM
    opts: opts_SFAM
    ...
julia> train!(art, x, y)